chore: remove obsolete job params and scripts

+++ jobs/ci-run/scripts/package-juju-source.sh
@@ -6,8 +6,14 @@ if [ -z "$JUJU_SOURCE_CHECKOUT" ]; then
     exit 1
 fi
 
+if [ -z "$GOVERSION" ]; then
+    echo "Must specify GOVERSION"
+    exit 1
+fi
+
 export PATH="/snap/bin:$PATH"
-sudo snap refresh go --channel=1.21/stable || sudo snap install go --channel=1.21/stable --classic
+GO_MAJOR_MINOR=$(echo ${GOVERSION} | cut -d '.' -f1,2)
+sudo snap refresh go --channel="${GO_MAJOR_MINOR}/stable" || sudo snap install go --channel="${GO_MAJOR_MINOR}/stable" --classic
 
 # TODO - fix this workaround
 # As we clone the full history (can't shallow clone otherwise queued jobs miss
@@ -22,8 +28,8 @@ full_path=${GOPATH}/src/github.com/juju/juju
 export PATH=/snap/bin:$PATH:$GOPATH/bin
 
 rm -rf ${full_path}
-git clone --depth 1 file://${JUJU_SOURCE_CHECKOUT} ${full_path}
-rm -fr ${JUJU_SOURCE_CHECKOUT}
+git clone --depth 1 "file://${JUJU_SOURCE_CHECKOUT}" "${full_path}"
+rm -fr "${JUJU_SOURCE_CHECKOUT}"
 
 # TODO - remove vendor mode for go mod
 # It currently is needed for s390x builds since the s390x slave
@@ -31,36 +37,35 @@ rm -fr ${JUJU_SOURCE_CHECKOUT}
 echo "Resolving dependencies"
 export JUJU_MAKE_DEP=true
 export JUJU_GOMOD_MODE=vendor
-make -C ${full_path} godeps || make -C ${full_path} dep || make -C ${full_path} vendor-dependencies
+make -C "${full_path}" godeps || make -C "${full_path}" dep || make -C "${full_path}" vendor-dependencies
 
 echo "Removing non-free data."
-rm -rf ${GOPATH}/src/github.com/rogpeppe/godeps
-rm -rf ${GOPATH}/src/github.com/golang/dep
-rm -rf ${GOPATH}/src/github.com/kisielk
-rm -rf ${GOPATH}/src/code.google.com/p/go.net/html/charset/testdata/
-rm -f ${GOPATH}/src/code.google.com/p/go.net/html/charset/*test.go
-rm -rf ${GOPATH}/src/golang.org/x/net/html/charset/testdata/
-rm -f ${GOPATH}/src/golang.org/x/net/html/charset/*test.go
-rm -rf ${GOPATH}/src/github.com/prometheus/procfs/fixtures
+rm -rf "${GOPATH}/src/github.com/rogpeppe/godeps"
+rm -rf "${GOPATH}/src/github.com/golang/dep"
+rm -rf "${GOPATH}/src/github.com/kisielk"
+rm -rf "${GOPATH}/src/code.google.com/p/go.net/html/charset/testdata"
+rm -f "${GOPATH}/src/code.google.com/p/go.net/html/charset/*test.go"
+rm -rf "${GOPATH}/src/golang.org/x/net/html/charset/testdata"
+rm -f "${GOPATH}/src/golang.org/x/net/html/charset/*test.go"
+rm -rf "${GOPATH}/src/github.com/prometheus/procfs/fixtures"
 
 # Remove backup files that confuse lintian.
 echo "Removing backup files"
-find ${GOPATH}/src/ -type f -name *.go.orig -delete
+find "${GOPATH}/src/" -type f -name "*.go.orig" -delete
 
 echo "Attempting to apply patches"
 (cd ${full_path} && GOPATH=${GOPATH} make add-patches || true)
 
 echo "Removing binaries and build artifacts"
 if [[ -d ${GOPATH}/bin ]]; then
-    rm -r ${GOPATH}/bin
+    rm -r "${GOPATH:?}"/bin
 fi
 if [[ -d ${GOPATH}/pkg ]]; then
     # go mod prevents writes
-    chmod +w -R ${GOPATH}/pkg
-    rm -r ${GOPATH}/pkg
+    chmod +w -R "${GOPATH}/pkg"
+    rm -r "${GOPATH}/pkg"
 fi
 
-PROPS_PATH=${WORKSPACE}/build.properties
 #  Need to prepare some variables used through out the process
 VERSION_FILE="${full_path}/version/version.go"
 if [ -f "${full_path}/core/version/version.go" ]; then
@@ -71,20 +76,24 @@ if [[ -n ${JUJU_BUILD_NUMBER:-} ]]; then
     JUJU_VERSION="${JUJU_VERSION}.${JUJU_BUILD_NUMBER}"
 fi
 
+PROPS_PATH=${WORKSPACE}/build.properties
 # GIT_COMMIT is used by juju Makefile for version info.
-GIT_COMMIT=$(git -C ${full_path} rev-parse HEAD)
+GIT_COMMIT=$(git -C "${full_path}" rev-parse HEAD)
 SHORT_GIT_COMMIT=${GIT_COMMIT:0:7}
-echo "SHORT_GIT_COMMIT=${SHORT_GIT_COMMIT}" >> $PROPS_PATH
-echo "GIT_COMMIT=${GIT_COMMIT}" >> $PROPS_PATH
-echo "GOVERSION=${GOVERSION}" >> $PROPS_PATH
-echo "JUJU_BUILD_NUMBER=${JUJU_BUILD_NUMBER}" >> $PROPS_PATH
-echo "JUJU_VERSION=${JUJU_VERSION}" >> $PROPS_PATH
-echo "JUJU_GOMOD_MODE=${JUJU_GOMOD_MODE}" >> $PROPS_PATH
-echo "JUJU_SRC_TARBALL=juju-source-${JUJU_VERSION}-${SHORT_GIT_COMMIT}.tar.xz" >> $PROPS_PATH
-echo "JUJU_VERSION_MAJOR_MINOR=$(echo ${JUJU_VERSION} | cut -d'-' -f 1 | cut -d'.' -f 1,2)" >> $PROPS_PATH
+cat > "${PROPS_PATH}" <<EOF
+SHORT_GIT_COMMIT=${SHORT_GIT_COMMIT}
+GIT_COMMIT=${GIT_COMMIT}
+GOVERSION=${GOVERSION}
+JUJU_BUILD_NUMBER=${JUJU_BUILD_NUMBER}
+JUJU_VERSION=${JUJU_VERSION}
+JUJU_GOMOD_MODE=${JUJU_GOMOD_MODE}
+JUJU_SRC_TARBALL=juju-source-${JUJU_VERSION}-${SHORT_GIT_COMMIT}.tar.xz
+JUJU_VERSION_MAJOR_MINOR=$(echo "${JUJU_VERSION}" | cut -d'-' -f 1 | cut -d'.' -f 1,2)
+EOF
 
-source ${WORKSPACE}/build.properties
+# shellcheck source=/dev/null
+source "${PROPS_PATH}"
 
-tar cfJ ${JUJU_SRC_TARBALL} \
+tar cfJ "${JUJU_SRC_TARBALL}" \
     --exclude .git --exclude .bzr --exclude .hg \
-    -C ${GOPATH} ./
+    -C "${GOPATH}" ./
